+++ package.json
@@ -9,12 +9,25 @@
     "web": "expo start --web"
   },
   "dependencies": {
-    "expo": "~48.0.6",
+    "@react-navigation/material-bottom-tabs": "^6.2.15",
+    "@react-navigation/material-top-tabs": "^6.6.2",
+    "@react-navigation/native": "^6.1.6",
+    "@react-navigation/stack": "^6.3.16",
+    "expo": "^48.0.6",
     "expo-print": "~12.2.1",
     "expo-sharing": "~11.2.2",
     "expo-status-bar": "~1.4.4",
+    "install": "^0.13.0",
+    "npx": "^10.2.2",
     "react": "18.2.0",
-    "react-native": "0.71.3"
+    "react-native": "0.71.3",
+    "react-native-gesture-handler": "~2.8.0",
+    "react-native-pager-view": "^6.1.4",
+    "react-native-paper": "^5.4.1",
+    "react-native-safe-area-context": "^4.5.0",
+    "react-native-screens": "^3.20.0",
+    "react-native-tab-view": "^3.5.1",
+    "react-native-vector-icons": "^9.2.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
